feat!: convert date to iso instead of number (#21)
The reason is that Date is an object, not a scalar value, so you can't have it serialized in JSON or Avro. The only way is to transform it to a Unix timestamp (number) or some string standard (like the ISO one).

We started with the numeric approach, but we realized that we prefer the ISO one since it's easier for humans to read.